Merge groupchat tour script from Scrappie
This allows users to sign up for a tournament via ``.signup FORMAT``.

The bot then creates a groupchat if none exists, and invites the user to
join.

A tournament will be created and advertised once, in whichever rooms the
bot is configured to post for that format. If the player cap is not
reached, the room will expire after 40 minutes.